requirements.txt
discord.py
openai
deepl